refactor dom names, join files
